Title: updates for use of new navbar * SVGImg.jsx: updates comments for the image component * NewBar.jsx (renamed): updates to use the svg component * App.jsx: updates indentation and new navbar as "default from NewBar.jsx"
