+++ fitness-app/frontend/src/components/Icons/SVGImg.jsx
@@ -3,16 +3,11 @@ import styles from './Icons.module.scss';
 /**
  * A component to use svg images as buttons
  * @param props props for the component:
- *
- * svgFile
- *
- * imgAlt
- *
- * styleClass
- *
- * handlerFunction
- *
- * stateValue
+ * @param props.svgFile the file containing the svg image data
+ * @param props.imgAlt the img alt-attribute
+ * @param props.styleClass the class for styling
+ * @param props.handlerFunction the react useState function
+ * @param props.stateValue optional useState value
  * @returns {JSX.Element}
  * @constructor
  */
